+++ src/img2proj.rs
@@ -1,26 +1,171 @@
 //! Module containing the structure to convert back on forth 
 //! from Image coordinates to Intermediate coordinates, i.e. coordinates in the projection plane.
 
+use std::ops::RangeInclusive;
 use crate::{
   ImgXY, ProjXY,
   sip::Sip
 };
 
+/// Transform the XY coordinates in the projection plane in a pixel coordinates in an image.
+pub trait ProjXY2ImgXY {
+  
+  /// Transforms intermediate world coordinates -- i.e. coordinates in the canonical projection 
+  /// plane -- to pixel coordinates in an image.
+  /// # Params
+  /// * `xy` coordinates in the canonical projection plane.
+  fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY>;
+}
+
+/// Transform the pixel coordinates in an image to the XY coordinates in the projection plane.
+pub trait ImgXY2ProjXY {
+  
+  /// Type of the inverse transformation.
+  type T: ProjXY2ImgXY;
+
+  /// Transforms pixel coordinates to the intermediate world coordinates, i.e.
+  ///  to coordinates in the canonical projection plane.
+  /// # Params
+  /// * `xy`: pixel coordinates (no units)
+  fn img2proj(&self, xy: &ImgXY) -> ProjXY;
+  
+  /// Provide the inverse transformation.
+  fn inverse(&self) -> Self::T;
+}
+
+#[derive(Clone)]
+pub struct BasicImgXY2ProjXY {
+  /// Projection x-axis coordinate at the image X-axis center.
+  center_px: f64,
+  /// Projection y-axis coordinate at the image Y-axis center.
+  center_py: f64,
+  /// Center of the image in the X-axis (= half the number of pixels along the X-axis).
+  center_x: f64,
+  /// Center of the image in the Y-axis (= half the number of pixels along the Y-axis).
+  center_y: f64,
+  /// Size of a pixel along the projection plane the x-axis.
+  scale_x: f64,
+  /// Size of a pixel along the projection plane the y-axis.
+  scale_y: f64,
+}
+
+impl BasicImgXY2ProjXY {
+  
+  /// We assume the origin to be on the bottom left corner of the image,
+  /// with the `x` increasing from left to right and `y` increasing bottom-up.
+  /// # Remark
+  /// * for a PNG, the origin is the top left corner and `y` is increasing top-down.
+  /// * for the east to be towards the left of an image, the `x` axis as to be reversed.
+  /// # Params
+  /// * `img_size`: `(size_x, size_y)` number of pixels in each dimension
+  /// * `proj_bounds`: `(bounds_x, bounds_y)` boundaries of the projection domain
+  pub fn from(
+    img_size: (u16, u16),
+    proj_bounds: (RangeInclusive<f64>, RangeInclusive<f64>)
+  ) -> Self {
+    // x-axis
+    let img_size_x = img_size.0 as f64;
+    let center_x = 0.5 * (img_size_x - 1.0);
+    let xp_min = proj_bounds.0.start();
+    let xp_max = proj_bounds.0.end();
+    let scale_x = (xp_max - xp_min) / img_size_x as f64;
+    let center_px = 0.5 * (xp_max + xp_min);
+    // y-axis
+    let img_size_y = img_size.1 as f64;
+    let center_y = 0.5 * (img_size_y - 1.0);
+    let yp_min = proj_bounds.1.start();
+    let yp_max = proj_bounds.1.end();
+    let scale_y = (yp_max - yp_min) / img_size_y as f64;
+    let center_py = 0.5 * (yp_max + yp_min);
+    // Create
+    Self {
+      center_px,
+      center_py,
+      center_x,
+      center_y,
+      scale_x,
+      scale_y
+    }
+  }
+}
+
+impl ImgXY2ProjXY for BasicImgXY2ProjXY {
+  type T = Self;
+
+  fn img2proj(&self, xy: &ImgXY) -> ProjXY {
+    let proj_x = self.center_px + (xy.x as f64 - self.center_x) * self.scale_x;
+    let proj_y = self.center_py + (xy.y as f64 - self.center_y) * self.scale_y;
+    ProjXY::new(proj_x, proj_y)
+  }
+
+  fn inverse(&self) ->  Self::T {
+    self.clone()
+  }
+}
+
+impl ProjXY2ImgXY for BasicImgXY2ProjXY {
+
+  fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY> {
+    let x = self.center_x + (xy.x - self.center_px) / self.scale_x;
+    let y = self.center_y + (xy.y - self.center_py) / self.scale_y;
+    Some(ImgXY::new(x, y))
+  }
+}
+
+
+/// Specific implementation for PNG (top left origin) with East towards the left.
+#[derive(Clone)]
+pub struct ReversedEastPngImgXY2ProjXY(BasicImgXY2ProjXY);
+
+impl ReversedEastPngImgXY2ProjXY {
+  pub fn from(
+    img_size: (u16, u16),
+    proj_bounds: (RangeInclusive<f64>, RangeInclusive<f64>)
+  ) -> Self {
+    Self(BasicImgXY2ProjXY::from(img_size, proj_bounds))
+  }
+}
+
+
+impl ImgXY2ProjXY for ReversedEastPngImgXY2ProjXY {
+  type T = Self;
+
+  fn img2proj(&self, xy: &ImgXY) -> ProjXY {
+    let proj_x = self.0.center_px + (xy.x as f64 - self.0.center_x) * self.0.scale_x;
+    let proj_y = self.0.center_py + (xy.y as f64 - self.0.center_y) * self.0.scale_y;
+    ProjXY::new(-proj_x, -proj_y)
+  }
+
+  fn inverse(&self) ->  Self::T {
+    self.clone()
+  }
+}
+
+impl ProjXY2ImgXY for ReversedEastPngImgXY2ProjXY {
+
+  fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY> {
+    let x = self.0.center_x + (-xy.x - self.0.center_px) / self.0.scale_x;
+    let y = self.0.center_y + (-xy.y - self.0.center_py) / self.0.scale_y;
+    Some(ImgXY::new(x, y))
+  }
+}
+
+
+
 /// Struct allowing to transform the pixel coordinates in an image to the XY coordinates 
 /// in the projection plane.
 /// The three constructors are each associated with one of the three convention
 /// describe in the FITS paper: CDij, CDELTi + PCij, CDELTi + CROTA2.
-pub struct ImgXY2ProjXY {
+#[derive(Clone)]
+pub struct WcsImgXY2ProjXY {
   /// Translation vector (in pixel units, so no units).
   crpix1: f64, crpix2: f64,
   /// Rotation (no units) combined with a scale (in radians) matrix.
   cd11: f64, cd12: f64,
   cd21: f64, cd22: f64,
-  /// Possible SIP transformation
-  sip: Option<Sip>,
 }
 
-impl ImgXY2ProjXY {
+impl WcsImgXY2ProjXY {
   
   /// Create a struct from the `CDij` convention.
   /// # Params
@@ -41,7 +186,6 @@ impl ImgXY2ProjXY {
       cd12: cd12.to_radians(),
       cd21: cd21.to_radians(),
       cd22: cd22.to_radians(),
-      sip: None
     }
   }
 
@@ -86,15 +230,11 @@ impl ImgXY2ProjXY {
     )
   }
   
-  /// Returnthe previous value, if any.
-  pub fn set_sip(&mut self, sip: Sip) -> Option<Sip> {
-    self.sip.replace(sip)
-  }
+}
 
-  /// Remove SIP component
-  pub fn rm_sip(&mut self) {
-    self.sip = None
-  }
+impl ImgXY2ProjXY for WcsImgXY2ProjXY {
+  
+  type T = WcsProjXY2ImgXY;
   
   /// Transform the pixel coordinates to the intermediate world coordinates
   /// (or native spherical coordinates) by applying first a translation
@@ -102,16 +242,10 @@ impl ImgXY2ProjXY {
   /// (given the `CDij` keywords values).
   /// # Params
   /// * `imgXY`: pixel coordinates (no units) to be transformed intermediate world coordinates
-  pub fn img2proj(&self, xy: &ImgXY) -> ProjXY {
+  fn img2proj(&self, xy: &ImgXY) -> ProjXY {
     // Translation
-    let mut x = xy.x - self.crpix1;
-    let mut y = xy.y - self.crpix2;
-    // Possible SIP convention
-    if let Some(sip) = &self.sip {
-      let tmp = x;
-      x += sip.f(x, y);
-      y += sip.g(tmp, y);
-    } 
+    let x = xy.x - self.crpix1;
+    let y = xy.y - self.crpix2;
     // Rotation + scale
     ProjXY::new(
       self.cd11 * x + self.cd12 * y,
@@ -119,44 +253,108 @@ impl ImgXY2ProjXY {
     )
   }
   
-  pub fn inverse(&self) -> ProjXY2ImgXY {
+  
+  fn inverse(&self) -> Self::T {
     // Compute the determinant of the CD matrix
     let det = self.cd11 * self.cd22 - self.cd12 * self.cd21;
     // Compute the coefficient of the inverse matrix
-    ProjXY2ImgXY {
-      crpix1: self.crpix1, 
+    WcsProjXY2ImgXY {
+      crpix1: self.crpix1,
       crpix2: self.crpix2,
       icd11:  self.cd22 / det,
       icd12: -self.cd21 / det,
       icd21: -self.cd12 / det,
       icd22:  self.cd11 / det,
-      sip: self.sip.clone()
     }
+  }
+}
+
+/// Struct allowing to transform the pixel coordinates in an image to the XY coordinates 
+/// in the projection plane.
+/// The three constructors are each associated with one of the three convention
+/// describe in the FITS paper: CDij, CDELTi + PCij, CDELTi + CROTA2.
+pub struct WcsWithSipImgXY2ProjXY {
+  /// Regular transformation
+  wcs: WcsImgXY2ProjXY,
+  /// SIP transformation
+  sip: Sip,
+}
+
+impl WcsWithSipImgXY2ProjXY {
+  
+  /// Add SIP convention to a regular WCS transformation.
+  pub fn new(wcs: WcsImgXY2ProjXY, sip: Sip) -> Self {
+    Self { wcs, sip }
   }
   
 }
 
-pub struct ProjXY2ImgXY {
+impl ImgXY2ProjXY for WcsWithSipImgXY2ProjXY {
+
+  type T = WcsWithSipProjXY2ImgXY;
+
+  /// Transform the pixel coordinates to the intermediate world coordinates
+  /// (or native spherical coordinates) by applying first a translation
+  /// (given the `CRPIXi` keywords value) and then a rotation plus a scale
+  /// (given the `CDij` keywords values).
+  /// # Params
+  /// * `imgXY`: pixel coordinates (no units) to be transformed intermediate world coordinates
+  fn img2proj(&self, xy: &ImgXY) -> ProjXY {
+    // Translation
+    let mut x = xy.x - self.wcs.crpix1;
+    let mut y = xy.y - self.wcs.crpix2;
+    let tmp = x;
+    x += self.sip.f(x, y);
+    y += self.sip.g(tmp, y);
+    // Rotation + scale
+    ProjXY::new(
+      self.wcs.cd11 * x + self.wcs.cd12 * y,
+      self.wcs.cd21 * x + self.wcs.cd22 * y
+    )
+  }
+  
+  fn inverse(&self) -> Self::T {
+    WcsWithSipProjXY2ImgXY {
+     wcs: self.wcs.inverse(),
+     sip: self.sip.clone()
+    }
+  }
+}
+
+
+
+pub struct WcsProjXY2ImgXY {
   /// Translation vector (in pixel units, so no units).
   crpix1: f64, crpix2: f64,
   /// Rotation (no units) combined with a scale (in radians) matrix.
   icd11: f64, icd12: f64,
   icd21: f64, icd22: f64,
-  /// Possible SIP transformation
-  sip: Option<Sip>,
 }
 
-impl ProjXY2ImgXY {
+impl ProjXY2ImgXY for WcsProjXY2ImgXY {
   
-  pub fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY> {
+  fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY> {
     let x = self.icd11 * xy.x + self.icd12 * xy.y + self.crpix1;
     let y = self.icd21 * xy.x + self.icd22 * xy.y + self.crpix2;
-    if let Some(sip) = &self.sip {
-      sip.inverse(x, y).map(|ImgXY{x: rx, y: ry}| ImgXY::new(x + rx, y + ry))
-    } else {
-      Some(ImgXY::new(x, y))
-    }
+    Some(ImgXY::new(x, y))
    }
+}
+
+
+pub struct WcsWithSipProjXY2ImgXY {
+  /// Regular transformation
+  wcs: WcsProjXY2ImgXY,
+  /// SIP transformation
+  sip: Sip,
+}
+
+impl ProjXY2ImgXY for WcsWithSipProjXY2ImgXY {
+
+  fn proj2img(&self, xy: &ProjXY) -> Option<ImgXY> {
+    let x = self.wcs.icd11 * xy.x + self.wcs.icd12 * xy.y + self.wcs.crpix1;
+    let y = self.wcs.icd21 * xy.x + self.wcs.icd22 * xy.y + self.wcs.crpix2;
+    self.sip.inverse(x, y).map(|ImgXY{x: rx, y: ry}| ImgXY::new(x + rx, y + ry))
+  }
 }